requirements.txt
# this is the "requirements.txt" file
requests
python-dotenv
pandas

#environment name stocks-env
#ALPHAVANTAGE_API_KEY="2LAMCMHNZLA9IEAS"